% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_kanton.R
\name{get_kanton}
\alias{get_kanton}
\title{Get the name, label or number of the canton from the Gemeindenummer}
\usage{
get_kanton(x, render = "label", bfs = TRUE)
}
\arguments{
\item{x}{a \code{vector} of codes to be converted}

\item{render}{Return \code{label} (default), the \code{name} or the \code{number}?}

\item{bfs}{\code{x} is the Gemeindenummer (default) or the canton number?}
}
\value{
a character vector of names/labels or a numeric vector.
}
\description{
Get the name, label or number of the canton from the Gemeindenummer
}
\details{
Each canton uses certain ranges of the Gemeindenummer (GEOSTAT-Nr. or BFS-Nr.)
which runs from 1 to 6806. This function converts Gemeindenummern either to the 
name of the canton (if \code{render="name"}), the two-character label 
(\code{render="label"}) or the canton number (\code{render="number"}). 

\code{x} can either be the 1-4 digit Gemeindenummer (set \code{bfs=TRUE})
or the 1-2 digits canton number (\code{bfs=FALSE}). If \code{x} is not 
a numeric vector, it is coerced to one.
}
\examples{
 \dontrun{
 
	 get_kanton( c("4951") )
  get_kanton( c("02", "3"), bfs=TRUE )

 }




}
